requirements.txt
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
distro==1.9.0
filelock==3.18.0
Flask==3.1.1
fsspec==2025.5.1
grpcio==1.72.1
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.32.4
hyperframe==6.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
MarkupSafe==3.0.2
mpmath==1.3.0
mysql-connector-python==9.3.0
networkx==3.5
numpy==2.2.6
openai==1.82.0
packaging==25.0
pillow==11.2.1
portalocker==2.10.1
protobuf==6.31.1
pydantic==2.11.5
pydantic_core==2.33.2
pytesseract==0.3.13
python-dotenv==1.1.0
pywin32==310
PyYAML==6.0.2
qdrant-client==1.14.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
sentence-transformers==4.1.0
setuptools==80.9.0
sniffio==1.3.1
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.7.0
tqdm==4.67.1
transformers==4.52.4
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
